Rewrite external scanner in C
