--- src/scanner.cc
@@ -1,69 +1,0 @@
-#include <tree_sitter/parser.h>
-
-enum TokenType {
-  NEWLINE,
-};
-
-struct Scanner {
-  Scanner() {
-    deserialize(NULL, 0);
-  }
-
-  unsigned serialize(char *buffer) {
-    return 0;
-  }
-
-  void deserialize(const char *buffer, unsigned length) {
-  }
-
-  void advance(TSLexer *lexer) {
-    lexer->advance(lexer, false);
-  }
-
-  void skip(TSLexer *lexer) {
-    lexer->advance(lexer, true);
-  }
-
-  bool scan(TSLexer *lexer, const bool *valid_symbols) {
-    if (valid_symbols[NEWLINE]) {
-      if (lexer->lookahead == '\n' || lexer->lookahead == 0) {
-        skip(lexer);
-        lexer->result_symbol = NEWLINE;
-        return true;
-      }
-    }
-    return false;
-  }
-};
-
-extern "C" {
-
-void *tree_sitter_lean_external_scanner_create() {
-  return new Scanner();
-}
-
-bool tree_sitter_lean_external_scanner_scan(void *payload, TSLexer *lexer,
-                                            const bool *valid_symbols) {
-  Scanner *scanner = static_cast<Scanner *>(payload);
-  return scanner->scan(lexer, valid_symbols);
-}
-
-unsigned tree_sitter_lean_external_scanner_serialize(void *payload,
-                                                     char *buffer) {
-  Scanner *scanner = static_cast<Scanner *>(payload);
-  return scanner->serialize(buffer);
-}
-
-void tree_sitter_lean_external_scanner_deserialize(void *payload,
-                                                   const char *buffer,
-                                                   unsigned length) {
-  Scanner *scanner = static_cast<Scanner *>(payload);
-  scanner->deserialize(buffer, length);
-}
-
-void tree_sitter_lean_external_scanner_destroy(void *payload) {
-  Scanner *scanner = static_cast<Scanner *>(payload);
-  delete scanner;
-}
-
-}
